Add LLVM dependency and improve tokenizer logic
- Integrate llvm-zig as a dependency and import its modules - Update
build scripts to add llvm and clang modules to the build - Enhance
tokenizer to handle identifiers and integers, and free memory - Change
Token Int type from u32 to u8 - Add compiler and parser source files -
Update main.zr to return 10 - Print tokens line by line in main.zig

+++ src/tokens.zig
@@ -5,7 +5,7 @@ pub const Token = struct {
 };
 
 pub const Type = union(enum) {
-    Int: u32,
+    Int: u8,
     Identifier: []const u8,
     LeftParenthesis,
     RightParenthesis,
